revert: remove broken 816-opt peephole optimizations
The following optimizations were added but produced incorrect code:
- sep/rep elimination (unsafe - removed necessary mode switches)
- push/pull pair elimination
- inc/dec cancellation

These optimizations caused sprite corruption in MetaSprite example
and likely other rendering issues. Reverting to original optimizer
until the optimizations can be properly validated.

+++ pvsneslib/include/snes/snestypes.h
@@ -43,10 +43,10 @@ typedef unsigned long long u32;
 
 typedef volatile unsigned char vuint8;
 typedef volatile unsigned short vuint16;
-typedef volatile unsigned long vuint32;
+typedef volatile unsigned long long vuint32;
 typedef volatile char vint8;
 typedef volatile short vint16;
-typedef volatile long vint32;
+typedef volatile long long vint32;
 
 // Handy function pointer typedefs
 //! a function pointer that takes no arguments and doesn't return anything.
